Extract Language preferences into its own sub-page
Move the Language form from the main Preferences screen into a dedicated
sub-page at /me/preferences/language. The main screen now shows a
RouterLinkSummaryButton link card with the current language as a badge.
